+++ seedfarmer/__main__.py
@@ -24,21 +24,11 @@ import seedfarmer
 from seedfarmer import DEBUG_LOGGING_FORMAT, commands, config, enable_debug
 from seedfarmer.cli_groups import bootstrap, init, list, remove, store
 from seedfarmer.output_utils import print_bolded
-from seedfarmer.services.session_manager import SessionManager
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
 # Load environment variables from .env file if it exists
 load_dotenv(dotenv_path=os.path.join(config.OPS_ROOT, ".env"), verbose=True, override=True)
-
-
-def _load_project() -> str:
-    try:
-        print_bolded("No --project provided, attempting load from seedfarmer.yaml", "white")
-        return config.PROJECT
-    except FileNotFoundError:
-        print_bolded("Unable to determine project to bootstrap, one of --project or a seedfarmer.yaml is required")
-        raise click.ClickException("Failed to determine project identifier")
 
 
 @click.group()
@@ -59,6 +49,18 @@ def version() -> None:
     required=True,
 )
 @click.option(
+    "--profile",
+    default=None,
+    help="The AWS profile to use for boto3.Sessions",
+    required=False,
+)
+@click.option(
+    "--region",
+    default=None,
+    help="The AWS region to use for boto3.Sessions",
+    required=False,
+)
+@click.option(
     "--debug/--no-debug",
     default=False,
     help="Enable detailed logging.",
@@ -76,7 +78,9 @@ def version() -> None:
     help="Write out the generated deployment manifest",
     show_default=True,
 )
-def apply(spec: str, debug: bool, dry_run: bool, show_manifest: bool) -> None:
+def apply(
+    spec: str, profile: Optional[str], region: Optional[str], debug: bool, dry_run: bool, show_manifest: bool
+) -> None:
     """Apply manifests to a SeedFarmer managed deployment"""
     if debug:
         enable_debug(format=DEBUG_LOGGING_FORMAT)
@@ -84,7 +88,9 @@ def apply(spec: str, debug: bool, dry_run: bool, show_manifest: bool) -> None:
     if dry_run:
         print_bolded(" ***   This is a dry-run...NO ACTIONS WILL BE TAKEN  *** ", "white")
 
-    commands.apply(spec, dry_run, show_manifest)
+    commands.apply(
+        deployment_manifest_path=spec, profile=profile, region_name=region, dryrun=dry_run, show_manifest=show_manifest
+    )
 
 
 @click.command(help="Destroy a SeedFarmer managed deployment")
@@ -106,13 +112,6 @@ def apply(spec: str, debug: bool, dry_run: bool, show_manifest: bool) -> None:
     show_default=True,
 )
 @click.option(
-    "--project",
-    "-p",
-    help="Project identifier",
-    required=False,
-    default=None,
-)
-@click.option(
     "--profile",
     default=None,
     help="The AWS profile to use for boto3.Sessions",
@@ -121,7 +120,7 @@ def apply(spec: str, debug: bool, dry_run: bool, show_manifest: bool) -> None:
 @click.option(
     "--region",
     default=None,
-    help="The AWS region to use for boto3.Sessions",
+    help="The AWS region to use for toolchain",
     required=False,
 )
 @click.option(
@@ -134,7 +133,6 @@ def destroy(
     deployment: str,
     dry_run: bool,
     show_manifest: bool,
-    project: Optional[str],
     profile: Optional[str],
     region: Optional[str],
     debug: bool,
@@ -142,16 +140,17 @@ def destroy(
     """Destroy a SeedFarmer managed deployment"""
     if debug:
         enable_debug(format=DEBUG_LOGGING_FORMAT)
-    if project is None:
-        project = _load_project()
+    # MUST use seedfarmer.yaml so we can initialize codeseeder configs
+    project = config.PROJECT
     _logger.debug("Listing all deployments for Project %s", project)
 
     _logger.info("Destroy for Project %s, Deployment %s", project, deployment)
     if dry_run:
         print_bolded(" ***   This is a dry-run...NO ACTIONS WILL BE TAKEN  *** ", "white")
 
-    (SessionManager().get_or_create(project_name=project, profile=profile, region_name=region).toolchain_session)
-    commands.destroy(deployment_name=deployment, dryrun=dry_run, show_manifest=show_manifest)
+    commands.destroy(
+        deployment_name=deployment, profile=profile, region_name=region, dryrun=dry_run, show_manifest=show_manifest
+    )
 
 
 def main() -> int: