+++ seedfarmer/cli_groups/_list_group.py
@@ -86,6 +86,18 @@ def list() -> None:
     show_default=True,
 )
 @click.option(
+    "--profile",
+    default=None,
+    help="The AWS profile to use for boto3.Sessions",
+    required=False,
+)
+@click.option(
+    "--region",
+    default=None,
+    help="The AWS region of the toolchain",
+    required=False,
+)
+@click.option(
     "--debug/--no-debug",
     default=False,
     help="Enable detailed logging.",
@@ -96,6 +108,8 @@ def list_deployspec(
     group: str,
     module: str,
     project: Optional[str],
+    profile: Optional[str],
+    region: Optional[str],
     target_account_id: Optional[str],
     target_region: Optional[str],
     debug: bool,
@@ -113,7 +127,7 @@ def list_deployspec(
     elif target_account_id is not None and target_region is not None:
         session = (
             SessionManager()
-            .get_or_create(project_name=project)
+            .get_or_create(project_name=project, profile=profile, region_name=region)
             .get_deployment_session(account_id=target_account_id, region_name=target_region)
         )
 
@@ -151,16 +165,16 @@ def list_deployspec(
     default=None,
 )
 @click.option(
-    "--target-account-id",
+    "--profile",
     default=None,
-    help="Account Id to remove module data from, if specifed --target-region is required",
-    show_default=True,
+    help="The AWS profile to use for boto3.Sessions",
+    required=False,
 )
 @click.option(
-    "--target-region",
+    "--region",
     default=None,
-    help="Region to remove module data from, if specifed --target-account-id is required",
-    show_default=True,
+    help="The AWS region of the toolchain",
+    required=False,
 )
 @click.option(
     "--export-local-env/--no-export-local-env",
@@ -179,8 +193,8 @@ def list_module_metadata(
     group: str,
     module: str,
     project: Optional[str],
-    target_account_id: Optional[str],
-    target_region: Optional[str],
+    profile: Optional[str],
+    region: Optional[str],
     export_local_env: str,
     debug: bool,
 ) -> None:
@@ -191,15 +205,13 @@ def list_module_metadata(
     if project is None:
         project = _load_project()
 
-    session: Optional[Session] = None
-    if (target_account_id is not None) != (target_region is not None):
-        raise ValueError("Must either specify both --target-account-id and --target-region, or neither")
-    elif target_account_id is not None and target_region is not None:
-        session = (
-            SessionManager()
-            .get_or_create(project_name=project)
-            .get_deployment_session(account_id=target_account_id, region_name=target_region)
-        )
+    session = SessionManager().get_or_create(project_name=project, profile=profile, region_name=region)
+    dep_manifest = du.generate_deployed_manifest(deployment_name=deployment, skip_deploy_spec=True)
+    dep_manifest.validate_and_set_module_defaults() if dep_manifest else None
+    session = session.get_deployment_session(
+        account_id=dep_manifest.get_module(group=group, module=module).get_target_account_id(),  # type: ignore
+        region_name=dep_manifest.get_module(group=group, module=module).target_region,  # type: ignore
+    )
 
     metadata_json = mi.get_module_metadata(deployment, group, module, session=session)
     if not export_local_env:
@@ -225,6 +237,18 @@ def list_module_metadata(
     required=True,
 )
 @click.option(
+    "--profile",
+    default=None,
+    help="The AWS profile to use for boto3.Sessions",
+    required=False,
+)
+@click.option(
+    "--region",
+    default=None,
+    help="The AWS region of the toolchain",
+    required=False,
+)
+@click.option(
     "--debug/--no-debug",
     default=False,
     help="Enable detailed logging.",
@@ -232,6 +256,8 @@ def list_module_metadata(
 )
 def list_modules(
     deployment: str,
+    profile: Optional[str],
+    region: Optional[str],
     debug: bool,
 ) -> None:
     if debug:
@@ -239,7 +265,7 @@ def list_modules(
     _logger.debug("We are getting modules for %s", deployment)
 
     project = _load_project()
-    SessionManager().get_or_create(project_name=project)
+    SessionManager().get_or_create(project_name=project, profile=profile, region_name=region)
 
     dep_manifest = du.generate_deployed_manifest(deployment_name=deployment, skip_deploy_spec=True)
     if dep_manifest:
@@ -263,7 +289,7 @@ def list_modules(
 @click.option(
     "--region",
     default=None,
-    help="The AWS region to use for boto3.Sessions",
+    help="The AWS region of the toolchain",
     required=False,
 )
 @click.option(